Mining works, bug fixes, compatibility improvements
This program is now capable of mining bitcoins. However, the mining pool
must be hard coded at the moment as user input is still not available
yet. Also, the program will appear to freeze when it's started up,
however, it is still mining bitcoins in the background. The freezing
needs to be fixed though as the screen will not update real time
information about the mining. That and user input still needs to be
fixed/added.